// src/collections/enums/field-type.enum.ts
export enum FieldType {
  TEXT = 'text',
  TEXTAREA = 'textarea',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  DATE = 'date',
  DATETIME = 'datetime',
  EMAIL = 'email',
  PASSWORD = 'password',
  URL = 'url',
  DROPDOWN_SINGLE = 'dropdown_single',
  DROPDOWN_MULTIPLE = 'dropdown_multiple',
  RELATION_ONE_TO_ONE = 'relation_one_to_one',
  RELATION_ONE_TO_MANY = 'relation_one_to_many',
  RELATION_MANY_TO_ONE = 'relation_many_to_one',
  RELATION_MANY_TO_MANY = 'relation_many_to_many',
  FILE = 'file',
  IMAGE = 'image',
  JSON = 'json',
  TRANSLATIONS = 'translations',
}